feat(capabilities): move asset operation under stellar capability (GET /stellar/asset)
