+++ apps/capabilities/src/capabilities/stellar/horizon.ts
@@ -11,7 +11,8 @@ export function isStellarAddress(value: string): boolean {
 }
 
 async function horizon(path: string): Promise<{ ok: boolean; status: number; data: unknown }> {
-  const res = await fetch(`${HORIZON_URL}${path}`, {
+  const baseUrl = process.env.STELLAR_HORIZON_URL ?? "https://horizon-testnet.stellar.org";
+  const res = await fetch(`${baseUrl}${path}`, {
     headers: { accept: "application/json" },
   });
   const data = await res.json().catch(() => null);
@@ -142,6 +143,60 @@ export async function getOrderbook(
   };
 }
 
+export interface AssetFlags {
+  authRequired: boolean;
+  authRevocable: boolean;
+  authImmutable: boolean;
+  authClawbackEnabled: boolean;
+}
+
+export interface AssetInfo {
+  code: string;
+  issuer: string;
+  amount: string;
+  numAccounts: number;
+  flags: AssetFlags;
+}
+
+/** Supply, holders count, and flags for an issued asset. */
+export async function getAssetInfo(code: string, issuer: string): Promise<AssetInfo> {
+  const path = `/assets?asset_code=${encodeURIComponent(code)}&asset_issuer=${encodeURIComponent(issuer)}`;
+  const { ok, data } = await horizon(path);
+  if (!ok) throw new Error("Asset query failed");
+
+  const records = (data as { _embedded?: { records?: HorizonAssetRecord[] } })?._embedded?.records;
+  if (!records || records.length === 0) {
+    throw new Error("Asset not found");
+  }
+
+  const r = records[0];
+  if (!r) {
+    throw new Error("Asset not found");
+  }
+
+  const amount = r.amount ?? r.balances?.authorized ?? "0";
+  const numAccounts =
+    r.num_accounts ??
+    (r.accounts
+      ? (r.accounts.authorized ?? 0) +
+        (r.accounts.authorized_to_maintain_liabilities ?? 0) +
+        (r.accounts.unauthorized ?? 0)
+      : 0);
+
+  return {
+    code: r.asset_code,
+    issuer: r.asset_issuer,
+    amount,
+    numAccounts,
+    flags: {
+      authRequired: Boolean(r.flags?.auth_required),
+      authRevocable: Boolean(r.flags?.auth_revocable),
+      authImmutable: Boolean(r.flags?.auth_immutable),
+      authClawbackEnabled: Boolean(r.flags?.auth_clawback_enabled),
+    },
+  };
+}
+
 interface HorizonBalance {
   balance: string;
   asset_type: string;
@@ -176,3 +231,24 @@ interface HorizonLedger {
   base_fee_in_stroops: number;
   base_reserve_in_stroops: number;
 }
+interface HorizonAssetRecord {
+  asset_code: string;
+  asset_issuer: string;
+  amount?: string;
+  num_accounts?: number;
+  balances?: {
+    authorized?: string;
+  };
+  accounts?: {
+    authorized?: number;
+    authorized_to_maintain_liabilities?: number;
+    unauthorized?: number;
+  };
+  flags?: {
+    auth_required?: boolean;
+    auth_revocable?: boolean;
+    auth_immutable?: boolean;
+    auth_clawback_enabled?: boolean;
+  };
+}
+
